feat(arrow-cast): move map cast tests to integration suite
